DATAGO-109064: allow spring message header key to solace user properties key mapping
